fix(ui): resolve usage analytics doctor findings
Lazy-load Recharts through the usage analytics chart surface so the dashboard avoids eager heavy imports while preserving Recharts composition. Simplify the weekly chart resize state, pre-index plan trends, and update the focused unit spec for lazy chart rendering.
